requirements.txt
streamlit
langchain==0.1.14
langchain-community==0.0.30
openai
python-dotenv
requests
